refactor(web): move share-page components under components/share/
The codebase namespaces all UI under apps/web/src/components/<domain>/
(sessions, dashboard, memories, meta, …); the share-page UI was the
only outlier living under app/s/[id]/_components/. Next.js does support
the underscore-prefix route-colocated pattern, but the codebase
convention is global components/, so conform.

  no-access.tsx          → components/share/no-access.tsx
  sign-in-to-view.tsx    → components/share/sign-in-to-view.tsx
  share-header-user.tsx  → components/share/header-user.tsx

The `share-` prefix on header-user is dropped now that the folder
encodes that namespace.
